Move RiskLevel and Key retrieval measurement in own class
