+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/transaction/RetrieveDiagnosisKeysTransaction.kt
@@ -200,13 +200,13 @@ object RetrieveDiagnosisKeysTransaction : Transaction() {
 
         if (CWADebug.isDebugBuildOrMode) {
             var totalFileSize: Long = 0
-            val fileSizeUnit = (1024 * 1024)
             keyFiles.also {
                 it.forEach { file ->
                     totalFileSize += file.length()
                 }
             }
-            onKeyFilesFinished?.invoke(keyFiles.size, totalFileSize / fileSizeUnit)
+            onKeyFilesFinished?.invoke(
+                keyFiles.size, totalFileSize)
             onKeyFilesFinished = null
         }
